add script to bootstrap db from fresh postgres instance
This commit adds a bash helper script which runs all of the .sql files
to create the GUTS database from a "fresh" postgres instance. It can be
used with a postgres instance on the same machine, or a hosted one,
assuming the hosted postgres instance allows connections from other
machines.