postgres: change submitted_at field from TIMESTAMP to VARCHAR
Initially, it seemed to be a good idea to use the TIMESTAMP WITH TIME
ZONE data type for this column. However, this data type is mostly useful
when it comes to doing operations with TIMESTAMP values within the
postgres engine itself.
It also stores the datetime string differently, omitting the timezone
from the string itself. The timezone is stored intrinsically in the
postgres engine, and is harder to access.

For the purposes of GUTS, storing the timestamp as a string is more than
appropriate.