Feat: Add spawner application
This commit adds a core component of GUTS; the spawner.

The spawner is, broadly, an application which waits for test requests,
and then spawns VMs for those tests to run on.

This commit comes complete with all unit testing of course.

More specifically, the main loop of the spawner does this, all while
periodically updating the last_updated_at field of the tests table:
- Checks for a "requested" test request with the highest priority
- sets the state of that test to "spawning"
- sets the vnc address
- checks whether the VM needs tpm, if the image is live, etc
- downloads the image
- creates a disk
- gets a qemu cmd line
- starts the vm
- waits for the test to finish in any way
- kills the vm
- removes the disk
- starts again

This commit also:
- modifies some database schema/permissions
- adds a couple more test files - a mini ISO and corresponding
  SHA256SUMS file